digital: Convert "sps" parameter from int to float in "correlate_and_sync_cc"
Even though the GRC XML file for correlate_and_sync_cc claimed sps
(samples per symbol) was a float, it was implemented in the block as an
unsigned int.  This commit fixes that, changing sps to a float.

+++ gr-digital/include/gnuradio/digital/correlate_and_sync_cc.h
@@ -77,7 +77,7 @@ namespace gr {
        *                  correlation (1.0). Default is 0.9.
        */
       static sptr make(const std::vector<gr_complex> &symbols,
-                       unsigned int sps, float threshold=0.9);
+                       float sps, float threshold=0.9);
 
       virtual std::vector<gr_complex> symbols() const = 0;
       virtual void set_symbols(const std::vector<gr_complex> &symbols) = 0;
